Remove scenario operation mode, rework API mode to enable arrival times and dependencies for "static" JSON-provided jobs

+++ .api/run_client.py
@@ -2,6 +2,8 @@
 
 import random
 import numpy
+import numpy.random
+from numpy.random import RandomState
 import time
 from scipy.stats import truncnorm
 from scipy.stats import lognorm
@@ -40,17 +42,20 @@ Represents a single job of a particular client.
 """
 class Job:
     
-    def __init__(self, user, name, filename, priority=1, wc_timeout=0, cpu_timeout=0):
+    def __init__(self, user, name, filename, priority=1, wc_timeout=0, cpu_timeout=0, arrival=0, dependencies=[]):
         self._user = user
         self._name = name
         self._filename = filename
         self._priority = priority
         self._wc_timeout = wc_timeout
         self._cpu_timeout = cpu_timeout
+        self._arrival = arrival
+        self._dependencies = dependencies
 
     def to_json(self):
-        return '{ "user": "%s", "name": "%s", "file": "%s", "priority": %.3f, "wallclock-limit": "%s", "cpu-limit": "%s" }' % (
-            self._user, self._name, self._filename, self._priority, self._wc_timeout, self._cpu_timeout)
+        d_str = str(self._dependencies).replace("'", "\"")
+        return '{ "user": "%s", "name": "%s", "file": "%s", "priority": %.3f, "wallclock-limit": "%s", "cpu-limit": "%s", "arrival": %.3f, "dependencies": %s }' % (
+            self._user, self._name, self._filename, self._priority, self._wc_timeout, self._cpu_timeout, self._arrival, d_str)
     
     def get_json_filename(self):
         return self._user + "." + self._name + ".json"
@@ -67,9 +72,11 @@ class Client:
         self._stream = []
         self._running_job_id = 1
     
-    def add_job_to_stream(self, filename, priority=1, wc_timeout=0, cpu_timeout=0):
-        self._stream += [Job(self._name, "job-" + str(self._running_job_id), filename, 
-                        priority=self._priority*priority, wc_timeout=wc_timeout, cpu_timeout=cpu_timeout)]
+    def add_job_to_stream(self, filename, priority=1, wc_timeout=0, cpu_timeout=0, arrival=0):
+        dependencies = [self._stream[-1]._user+"."+self._stream[-1]._name] if self._stream else []
+        self._stream += [Job(self._name, "job-" + str(self._running_job_id), filename,
+                        priority=self._priority*priority, wc_timeout=wc_timeout, cpu_timeout=cpu_timeout, 
+                        arrival=max(arrival,self._arrival), dependencies=dependencies)]
         self._running_job_id += 1
     
     def has_next_job(self):
@@ -112,13 +119,15 @@ def log(msg, args=()):
 Returns a single random sample from a normal distribution truncated at lower and upper limits.
 """
 def get_truncated_normal(mean=0, sd=1, low=0, upp=10):
-    return int(truncnorm((low - mean) / sd, (upp - mean) / sd, loc=mean, scale=sd).rvs().round())
+    global random_state
+    return int(truncnorm((low - mean) / sd, (upp - mean) / sd, loc=mean, scale=sd).rvs(random_state=random_state).round())
 
 """
 Returns a single random sample from a log-normal distribution truncated at lower and upper limits.
 """
 def get_lognorm(s, low=1, upp=100):
-    val = lognorm(s).rvs()
+    global random_state
+    val = lognorm(s).rvs(random_state=random_state)
     return max(low, min(upp, int(val.round())))
 
 """
@@ -154,26 +163,29 @@ def introduce_job(job, options):
 
 def main():
     global global_starttime
+    global random_state
 
     usage = "usage: %prog [options] <benchmark_file>"
     parser = OptionParser(usage=usage)
     
-    parser.add_option("-n", "--num-clients",          type="int",   dest="num_clients",          default="1",   help="number of clients arriving")
-    parser.add_option("-i", "--interarrival-time",    type="float", dest="interarrival_time",    default="10",  help="interarrival time of clients")
-    parser.add_option("-p", "--num-client-processes", type="int",   dest="num_client_processes", default="1",   help="number of client processes operating mallob's API")
+    parser.add_option("-n", "--num-clients",          type="int",   dest="num_clients",          default="1",     help="number of clients arriving")
+    parser.add_option("-i", "--interarrival-time",    type="float", dest="interarrival_time",    default="10",    help="interarrival time of clients")
+    parser.add_option("-p", "--num-client-processes", type="int",   dest="num_client_processes", default="1",     help="number of client processes operating mallob's API")
     
-    parser.add_option("-m", "--min-stream-length",    type="int",   dest="min_stream_length",    default="1",   help="minimum length of any job stream")
-    parser.add_option("-x", "--max-stream-length",    type="int",   dest="max_stream_length",    default="100", help="maximum length of any job stream")
-    parser.add_option("-e", "--mean-stream-length",   type="float", dest="mean_stream_length",   default="1",   help="mean length of job streams")
-    parser.add_option("-d", "--stddev-stream-length", type="float", dest="stddev_stream_length", default="3",   help="standard deviation of length of job streams")
+    parser.add_option("-m", "--min-stream-length",    type="int",   dest="min_stream_length",    default="1",     help="minimum length of any job stream")
+    parser.add_option("-x", "--max-stream-length",    type="int",   dest="max_stream_length",    default="100",   help="maximum length of any job stream")
+    parser.add_option("-e", "--mean-stream-length",   type="float", dest="mean_stream_length",   default="1",     help="mean length of job streams")
+    parser.add_option("-d", "--stddev-stream-length", type="float", dest="stddev_stream_length", default="3",     help="standard deviation of length of job streams")
     
-    parser.add_option("-M", "--min-stream-width",    type="int",   dest="min_stream_width",      default="1",   help="minimum width of any job stream")
-    parser.add_option("-X", "--max-stream-width",    type="int",   dest="max_stream_width",      default="1",   help="maximum width of any job stream")
-    parser.add_option("-S", "--shape-stream-width",  type="float", dest="shape_stream_width",    default="1",   help="shape parameter for log-normal width of job streams")
+    parser.add_option("-M", "--min-stream-width",     type="int",   dest="min_stream_width",     default="1",     help="minimum width of any job stream")
+    parser.add_option("-X", "--max-stream-width",     type="int",   dest="max_stream_width",     default="1",     help="maximum width of any job stream")
+    parser.add_option("-S", "--shape-stream-width",   type="float", dest="shape_stream_width",   default="1",     help="shape parameter for log-normal width of job streams")
     
-    parser.add_option("-s", "--seed",                 type="float", dest="seed",                 default="0",   help="random seed (0 for no seed)")
-    parser.add_option("-w", "--wallclock-limit",      type="float", dest="wc_limit_per_job",     default="0",   help="wallclock limit per job in seconds (0 for no limit)")
-    parser.add_option("-c", "--cpu-limit",            type="float", dest="cpu_limit_per_job",    default="0",   help="cpu limit per job in CPU seconds (0 for no limit)")
+    parser.add_option("-s", "--seed",                 type="float", dest="seed",                 default="0",     help="random seed (0 for no seed)")
+    parser.add_option("-w", "--wallclock-limit",      type="float", dest="wc_limit_per_job",     default="0",     help="wallclock limit per job in seconds (0 for no limit)")
+    parser.add_option("-c", "--cpu-limit",            type="float", dest="cpu_limit_per_job",    default="0",     help="cpu limit per job in CPU seconds (0 for no limit)")
+
+    parser.add_option("-I", "--instant-output", action="store_true", dest="instant_output",      default="False", help="Instantly output all job files with their arrival time")
     
     (options, args) = parser.parse_args()
     if len(args) != 1:
@@ -184,6 +196,9 @@ def main():
     
     if options.seed != 0:
         random.seed(options.seed)
+        random_state = RandomState(int(1000*options.seed))
+    else:
+        random_state = RandomState()
 
     # Create a number of random clients with a random job stream each
     log("Creating clients ...")
@@ -199,17 +214,20 @@ def main():
 
     # Remember the currently active job for each client
     active_jobs = [None for c in clients]
-    log("Clients set up. Waiting for mallob ...")
+    log("Clients set up.")
 
-    # Wait for mallob process
-    mallob_active = False
-    while not mallob_active:
-        for process in psutil.process_iter():
-            if process.name() == "mallob":
-                mallob_active = True
-                break
-        time.sleep(0.1)
-    log("Process named \"mallob\" is active!")
+    if not options.instant_output:
+        # Wait for mallob process
+        log("Waiting for mallob ...")
+        mallob_active = False
+        while not mallob_active:
+            for process in psutil.process_iter():
+                if process.name() == "mallob":
+                    mallob_active = True
+                    break
+            time.sleep(0.1)
+        log("Process named \"mallob\" is active!")
+
     global_starttime = time.time_ns()
     log("Launching clients.")
 
@@ -217,7 +235,7 @@ def main():
     any_left = True
     while any_left:
         time.sleep(0.1)
-        elapsed = elapsed_time()
+        elapsed = float("inf") if options.instant_output else elapsed_time()
         any_left = False
         
         # For each client
@@ -238,29 +256,39 @@ def main():
                 any_left = True
             
             elif active_jobs[i] is not None:
-                # Check if job finished
-                done_file = ".api/jobs.0/done/" + active_jobs[i].get_json_filename()
-                if os.path.isfile(done_file):
-                    # -- job finished
+                if options.instant_output:
+                    # Introduce next job directly
+                    # Introduce next job
+                    if c.has_next_job():
+                        active_jobs[i] = c.get_next_job()
+                        introduce_job(active_jobs[i], options)
+                    else:
+                        log("%s completed", (c._name,))
+                        active_jobs[i] = None
+                else:
+                    # Check if job finished
+                    done_file = ".api/jobs.0/done/" + active_jobs[i].get_json_filename()
+                    if os.path.isfile(done_file):
+                        # -- job finished
 
-                    f = open(done_file, 'r')
-                    try:
-                        j = json.load(f)
-                        log("%s finished (response time: %.3f, result code: %i)", (c._name + "." + active_jobs[i]._name, j["result"]["responsetime"], j["result"]["resultcode"]))
-                        f.close()
-                        os.remove(done_file)
-                        
-                        # Introduce next job
-                        if c.has_next_job():
-                            active_jobs[i] = c.get_next_job()
-                            introduce_job(active_jobs[i], options)
-                        else:
-                            log("%s completed", (c._name,))
-                            active_jobs[i] = None
+                        f = open(done_file, 'r')
+                        try:
+                            j = json.load(f)
+                            log("%s finished (response time: %.3f, result code: %i)", (c._name + "." + active_jobs[i]._name, j["result"]["responsetime"], j["result"]["resultcode"]))
+                            f.close()
+                            os.remove(done_file)
+                            
+                            # Introduce next job
+                            if c.has_next_job():
+                                active_jobs[i] = c.get_next_job()
+                                introduce_job(active_jobs[i], options)
+                            else:
+                                log("%s completed", (c._name,))
+                                active_jobs[i] = None
 
-                    except json.decoder.JSONDecodeError as e:
-                        # Could not parse JSON file - probably it is still being written. Try again later
-                        f.close()
+                        except json.decoder.JSONDecodeError:
+                            # Could not parse JSON file - probably it is still being written. Try again later
+                            f.close()
 
     log("Done.")
